Rename things and use Dialog instead of Drawer
